+++ .sequelizerc
@@ -3,9 +3,9 @@ require("@babel/register");
 const path = require('path');
 
 module.exports = {
-  "config": path.resolve('./server/config', 'config.js'),
-  "models-path": path.resolve('./server/models'),
-  "seeders-path": path.resolve('./server/seeders'),
-  "migrations-path": path.resolve('./server/migrations')
+  "config": path.resolve('./src/config', 'config.js'),
+  "models-path": path.resolve('./src/db/models'),
+  "seeders-path": path.resolve('./src/db/seeders'),
+  "migrations-path": path.resolve('./src/db/migrations')
 
 };
