--- src/config/index.js
@@ -1,4 +1,0 @@
-module.exports = {
-    secret:
-        process.env.NODE_ENV === "production" ? process.env.SECRET : "secret"
-};
